Rename history_import to history_link
