+++ client/src/components/Markdown/Markdown.vue
@@ -44,7 +44,7 @@
                 <div v-else-if="obj.name == 'generate_time'" class="galaxy-time">
                     <pre><code>{{ getTime }}</code></pre>
                 </div>
-                <HistoryImport v-else-if="obj.name == 'history_import'" :args="obj.args" :histories="histories" />
+                <HistoryLink v-else-if="obj.name == 'history_link'" :args="obj.args" :histories="histories" />
                 <HistoryDatasetAsImage v-else-if="obj.name == 'history_dataset_as_image'" :args="obj.args" />
                 <HistoryDatasetLink v-else-if="obj.name == 'history_dataset_link'" :args="obj.args" />
                 <HistoryDatasetIndex v-else-if="obj.name == 'history_dataset_index'" :args="obj.args" />
@@ -106,7 +106,7 @@ import HistoryDatasetLink from "./Elements/HistoryDatasetLink";
 import HistoryDatasetIndex from "./Elements/HistoryDatasetIndex";
 import HistoryDatasetCollectionDisplay from "./Elements/HistoryDatasetCollection/CollectionDisplay";
 import HistoryDatasetDetails from "./Elements/HistoryDatasetDetails";
-import HistoryImport from "./Elements/HistoryImport";
+import HistoryLink from "./Elements/HistoryLink";
 import InvocationTime from "./Elements/InvocationTime";
 import JobMetrics from "./Elements/JobMetrics";
 import JobParameters from "./Elements/JobParameters";
@@ -139,7 +139,7 @@ export default {
         HistoryDatasetDisplay,
         HistoryDatasetIndex,
         HistoryDatasetLink,
-        HistoryImport,
+        HistoryLink,
         JobMetrics,
         JobParameters,
         LoadingSpan,
